requirements.txt
requests>=2.31.0
python-telegram-bot>=20.7
python-dotenv>=1.0.0
undetected-chromedriver>=3.5.4
selenium>=4.16.0
